sort: allow $status_format %s expando to be conditional
The value of $sort_aux matters even when $sort is not "threads", but
only for breaking ties.  But how often do you have ties for
$sort=date?  (Okay, ties on $sort=from are more likely).  More
importantly, an upcoming patch wants to add a config variable that
changes inter-thread sorting differently than intra-thread sorting;
while $sort_aux will remain valid for breaking ties even when $sort is
not threads, the choice for inter-thread sorting makes no sense except
when $sort=threads.  Thus, it makes sense to be able to conditionalize
the status line on whether $sort is threads or anything else.  The
easiest way to do this is to special-case the %?s?...? conditional to
choose the else-part when %s is threads (just like we choose the
else-part when numeric values are 0).