sort: add $sort_thread config variable
Fixes: https://github.com/neomutt/neomutt/issues/2342

Add a new config variable $sort_thread, which makes the sorting of
top-level threads orthogonal from the sorting of subthreads within a
thread.  This allows us to finally achieve parity with Thunderbird,
where a thread with the most recent activity (at any depth) is floated
to the top or bottom of the list (sort_thread=reverse-last-date or
last-date), but within the thread, the subthreads are sorted by the
date that sibling subthreads started and not by when out-of-order
replies to those subthreads were added (sort_aux=date).  The new
config variable is only useful when $sort=threads; otherwise, it is
not consulted.  Furthermore, when at its default value of threads, it
defers to $sort_aux (for backwards compatibility); in fact, use of
last-threads or reverse-last-threads is an easy way to dynamically
track changes to sort_aux on determining which thread to move to the
desired end.

Implementation-wise, each thread now has to track two sort keys: the
message to use when sorting intra-thread siblings, and the message to
use when sorting between two overall threads.  While we are still able
to implement sort_aux=last-* with an O(1) peek at the first or last
sibling, properly implementing sort_thread=last-* requires an O(n)
pass over all siblings (although our runtime is still dominated by the
O(n lg n) nature of sorting in general).